created a new person api route
